requirements.txt
typing-extensions>=4.5.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
httpx>=0.25.0
openai>=1.10.0
jupyterlab>=4.0.0
anyio>=4.7.0
mypy>=1.13.0
pytest>=8.3.5
ruff>=0.8.2
python-dotenv>=1.0.1
psycopg
psycopg-pool
datasets

langchain-openai>=0.0.5
langgraph-checkpoint-postgres
langgraph-prebuilt>=0.1.8
langgraph>=0.3.21
langgraph-cli[inmem]>=0.2.8
langchain>=0.3.19
langchain-core>=0.3.39
langchain-community>=0.3.18

langchain-mcp-adapters
langchain_qdrant
